+++ deploy/tauri/src-tauri/i18n/en-US.json
@@ -1,3 +1,7 @@
 {
-  "hiden_to_tray": "hiden to tray ♬"
+  "hiden_to_tray": "hiden to tray ♬",
+  "tray_open": "Goto Read",
+  "tray_logs": "View Logs",
+  "tray_restart": "Restart Backend",
+  "tray_quit": "Quit"
 }
